Obtain expression matrix for degraded transcripts

This function is used to obtain a RangedSummarizedExperiment-class of transcripts and their expression values #' These transcripts are selected based on a prior study of RNA degradation in postmortem brain tissues. This object can later be used to obtain the principle components necessary to remove the effect of degradation in differential expression.

getDegTx(
  rse_tx,
  type = c("cell_component", "standard", "top1500"),
  sig_transcripts = select_transcripts(type),
  assayname = "tpm"
)

Arguments

rse_tx

A RangedSummarizedExperiment-class object containing the transcript data desired to be studied.

type

A character(1) specifying the transcripts set type. These were determined by Joshua M. Stolz et al, 2022. Here the names "cell_component", "top1500", and "standard" refer to models that were determined to be effective in removing degradation effects. The "standard" model involves taking the union of the top 1000 transcripts associated with degradation from the interaction model and the main effect model. The "top1500" model is the same as the "standard model except the union of the top 1500 genes associated with degradation is selected. The most effective of our models, "cell_component", involved deconvolution of the degradation matrix to determine the proportion of cell types within our studied tissue. These proportions were then added to our model.matrix() and the union of the top 1000 transcripts in the interaction model, the main effect model, and the cell proportions model were used to generate this model of qSVs.

sig_transcripts

A list of transcripts determined to have degradation signal in the qsva expanded paper.

assayname

character string specifying the name of the assay desired in rse_tx

Value

A RangedSummarizedExperiment-class

object.

Examples

getDegTx(covComb_tx_deg)
#> class: RangedSummarizedExperiment 
#> dim: 2976 60 
#> metadata(0):
#> assays(1): tpm
#> rownames(2976): ENST00000442987.3 ENST00000623083.3 ...
#>   ENST00000361681.2 ENST00000361789.2
#> rowData names(22): source type ... protein_id ccdsid
#> colnames(60): DLPFC_Br1385_A1_ribo DLPFC_Br1385_A2_ribo ... Br5941-C5c
#>   Br5941-C5d
#> colData names(21): SAMPLE_ID RNum ... totalAssignedGene totalMapped
stopifnot(mean(rowMeans(assays(covComb_tx_deg)$tpm)) > 1)