Funtion to obtain expression matrix for degraded transcripts

This function is used to obtain a RangedSummarizedExperiment-class of transcripts and their expression values #' These transcripts are selected based on a prior study of RNA degradation in postmortem brain tissues. This object can later be used to obtain the principle components necessary to remove the effect of degradation in differential expression.

getDegTx(rse_tx, sig_transcripts)

Arguments

rse_tx

A RangedSummarizedExperiment-class object containing the transcript data desired to be studied.

sig_transcripts

A list of transcripts determined to haave degradation signal in the qsva expanded paper.

Value

A RangedSummarizedExperiment-classobject.

Examples

getDegTx(covComb_tx_deg, rownames(covComb_tx_deg))
#> class: RangedSummarizedExperiment 
#> dim: 2976 119 
#> metadata(0):
#> assays(1): tpm
#> rownames(2976): ENST00000442987.3 ENST00000623083.3 ...
#>   ENST00000361681.2 ENST00000361789.2
#> rowData names(22): source type ... protein_id ccdsid
#> colnames(119): DLPFC_Br1385_A1_ribo DLPFC_Br1385_A2_ribo ...
#>   R15702_HJ7YKBBXX R15703_HJ7YKBBXX
#> colData names(21): SAMPLE_ID RNum ... totalAssignedGene totalMapped