A wrapper function used to perform qSVA in one step.

A wrapper function used to perform qSVA in one step.

qSVA(
  rse_tx,
  type = "cell_component",
  sig_transcripts = select_transcripts(type),
  mod,
  assayname
)

Arguments

rse_tx

A RangedSummarizedExperiment-class object containing the transcript data desired to be studied.

type

a character string specifying which model you would like to use when selecting a degradation matrix.

sig_transcripts

A list of transcripts determined to have degradation signal in the qsva expanded paper.

mod

Model Matrix with necessary variables the you would model for in differential expression

assayname

character string specifying the name of the assay desired in rse_tx

Value

Examples

mod <- model.matrix(~ mitoRate + Region + rRNA_rate + totalAssignedGene + RIN, data = colData(covComb_tx_deg))
qSVA(rse_tx = covComb_tx_deg, type = "cell_component", mod = mod, assayname = "tpm")
#>                              qSV1        qSV2        qSV3         qSV4
#> DLPFC_Br1385_A1_ribo -17.33886602  -7.7274895  21.5403997  -2.48331741
#> DLPFC_Br1385_A2_ribo   6.15454944 -13.1810261  13.1364975  -2.90972482
#> DLPFC_Br1385_A3_ribo  37.78999694 -20.4563897   5.3098690  -4.59091223
#> DLPFC_Br1385_A4_ribo  25.46489182 -15.7216611  15.4841808  -4.39926517
#> DLPFC_Br1729_B1_ribo -21.28542455  -3.8898501  25.6093521   1.11814070
#> DLPFC_Br1729_B2_ribo -11.51484495  -4.6040947  23.9912311   0.17288166
#> DLPFC_Br1729_B3_ribo  48.67861719 -18.2190098  15.5670867   2.23343372
#> DLPFC_Br1729_B4_ribo  14.29861049  -9.1939507  26.0130889   1.88708118
#> DLPFC_Br2015_F1_ribo -25.82779949  -3.0207869   9.7167808  -4.60608583
#> DLPFC_Br2015_F2_ribo  -2.23144157  -5.5394776  11.1131593  -8.15627934
#> DLPFC_Br2015_F3_ribo  18.28229280 -15.8710287   9.1096015  -0.48615888
#> DLPFC_Br2015_F4_ribo  13.95103639 -15.8719994  18.8819234   0.18012580
#> DLPFC_Br2020_C1_ribo -20.24920857  -5.1373404  21.0893298  -0.68529389
#> DLPFC_Br2020_C2_ribo  10.75734230  -9.0232940  23.1929478  -0.94338144
#> DLPFC_Br2020_C3_ribo  58.50366838 -25.0403470  11.3558387  -0.13203688
#> DLPFC_Br2020_C4_ribo  32.72045643 -13.8332263  20.9018831  -0.35444967
#> DLPFC_Br2074_D1_ribo -40.83616999  -2.3624372  11.9590411   4.42200326
#> DLPFC_Br2074_D2_ribo -15.63316696  -5.5792888  25.8198020   4.16586619
#> DLPFC_Br2074_D3_ribo  30.82783925 -21.2831725  25.4471922   6.92126867
#> DLPFC_Br2074_D4_ribo  -5.45361395  -7.7789567  24.1935374   5.05328295
#> Br2623-H1a           -27.35933819  -4.8653284  -5.3994055  -0.78037476
#> Br2623-H1b           -12.81673182  -4.5949898  -3.3551544  -4.81469308
#> Br2623-H1c           -13.49825960  -7.4562103  -1.9195944   0.63763120
#> Br2623-H1d             5.46491482 -11.3461904  -7.5647119  -1.53456400
#> Br5005-H2a           -34.47409117   8.5436977   2.2572622   2.62829289
#> Br5005-H2b            -4.43745796  13.2942539   4.0144359  -3.01975063
#> Br5005-H2c           -14.65469442   4.6442547  -1.1006187  -2.65773597
#> Br5005-H2d            -2.75136771   5.3695034  -4.5740263  -0.77711332
#> Br5370-H3a           -20.86407030   0.5172802  -6.1391223  18.22019470
#> Br5370-H3b             2.06036216   2.7951842  -3.4979698  16.57197761
#> Br5370-H3c             5.24793382   2.0569071  -7.2805259  13.29447464
#> Br5370-H3d            20.41617406  -6.3638525 -14.7939038  18.32331363
#> Br5889-H4a             3.30665212  20.1122457   4.5548808   2.30351030
#> Br5889-H4b            11.15407006  23.4016564   3.4603500  -0.57895065
#> Br5889-H4c            29.19643188  21.8472568   0.4630798  -0.69559860
#> Br5889-H4d            33.58482074  19.3773906  -8.5805213  -3.78870956
#> Br5941-H5a           -32.65161301  -2.5839653  -6.4104647  -1.32130394
#> Br5941-H5b           -14.68110004  -4.1684058  -7.4462830  -3.50443416
#> Br5941-H5c            -3.15596080  -5.5951067  -2.7156452  -4.56811748
#> Br5941-H5d            -4.58391311  -5.9064106  -8.8647682  -3.31665320
#> Br2623-C1a            -7.40294362  -4.8000613  -4.1385740 -10.03716019
#> Br2623-C1b            -9.57634213   4.3262541  -9.8714986 -14.05888981
#> Br2623-C1c            -4.18895594  -5.1523136  -8.0480461 -11.83990784
#> Br2623-C1d             2.29106500  -4.7760236  -7.4330093 -12.26303062
#> Br5005-C2a           -14.05158950  13.5264168   3.3178823  -4.93337971
#> Br5005-C2b            -5.44579588  12.4072637   1.1432395  -7.66145463
#> Br5005-C2c            -6.80556670  11.8893823  -0.1373716  -6.76259242
#> Br5005-C2d             8.33133864  11.6760417  -2.0675381  -7.89640550
#> Br5370-C3a            14.24329828   5.2990242 -14.9481366  11.11117573
#> Br5370-C3b            22.93489965   7.8323598 -16.5253280   7.90482799
#> Br5370-C3c            29.01257473   2.9239177 -18.3392298  10.65232857
#> Br5370-C3d            31.52056955  -1.5762517 -19.3907156   9.90554218
#> Br5889-C4a             3.75157452  30.9954063   8.1507415  -5.96608256
#> Br5889-C4b            17.01695849  36.6522358   5.5642607 -10.65424071
#> Br5889-C4c            26.13465411  36.1449670   1.1288989 -10.60349710
#> Br5889-C4d            31.00230541  33.6908415  -2.9161280 -12.29552817
#> Br5941-C5a            -0.51577007  -1.4660821  -1.1469090 -11.98199785
#> Br5941-C5b            -4.04368133   4.1180724  -3.9900324 -12.08787641
#> Br5941-C5c             0.70589034  -1.5033888  -4.2099753 -11.79882236
#> Br5941-C5d             7.72856069  -3.5880811  -3.5330538 -11.64100460
#> Br2623-M1a           -15.48516460  -9.7976071   0.9227947  -0.29282353
#> Br2623-M1b           -10.85303991  -9.0033807   4.1666081  -0.03366095
#> Br2623-M1c           -14.87925867  -9.2370000   4.6280496  -1.61602919
#> Br2623-M1d            -0.09262256  -9.9908398   2.0793343  -0.72453659
#> Br5005-M2a           -10.82372483   7.6850214   5.9412818  -0.42477168
#> Br5005-M2b           -15.97227060   7.0621760   5.5942213   0.97248556
#> Br5005-M2d            11.62699195   3.0513009   7.3250369   1.39784360
#> Br5370-M3a            -1.78036793  -0.7364666   1.9812623  21.22130679
#> Br5370-M3b             7.57527735   0.8665392  -0.3891052  19.27709838
#> Br5370-M3c            17.26751279  -0.4609492  -4.4695036  19.20743732
#> Br5370-M3d            28.97262041  -2.1731349  -7.4984501  16.19695689
#> Br5889-M4a            -4.33417250  23.0455216   8.8882500   2.58834649
#> Br5889-M4b             7.93821770  21.5700275  14.5183363   2.59340958
#> Br5889-M4c            23.76543685  21.1643586  13.0037787   3.14238644
#> Br5889-M4d            37.82923933  18.3990252   3.3242483   1.73134708
#> Br5941-M5a            -8.91668335  -4.2770910   1.6452934  -6.17316384
#> Br5941-M5b           -10.08232301  -5.0431059   1.6552003  -4.60238460
#> Br5941-M5c            -3.05889273  -6.2341614   3.8646694  -3.34153895
#> Br5941-M5d             6.84330936  -9.1915736   1.1399669  -2.80053902
#> R15684_HJ7YKBBXX     -38.37242183  -3.7840189  -4.6567710  20.62362484
#> R15685_HJ7YKBBXX     -15.98407812  -5.4249559  -2.6239270  23.00437198
#> R15686_HJ7YKBBXX      -0.36072004  -5.6739745  -4.2813045  23.30445993
#> R15687_HJ7YKBBXX      19.03181580 -10.6452777 -10.3347900  21.24445571
#> R15688_HJ7YKBBXX     -39.19691715  10.1025703  -1.9365401   2.00692534
#> R15689_HJ7YKBBXX     -32.69886352   8.4160164  -2.4492879   1.97641188
#> R15690_HJ7YKBBXX     -13.74553787   5.9570899  -7.1498873  -1.32258073
#> R15691_HJ7YKBBXX       7.89456850   3.4402467  -5.3283528   1.69195938
#> R15692_HJ7YKBBXX     -43.06342025  18.5658296   5.2930949   5.92999730
#> R15693_HJ7YKBBXX      -6.99564946  17.5843656   7.4786559   4.69816251
#> R15694_HJ7YKBBXX      -4.87654209  17.7409532   8.9718673   4.01895003
#> R15695_HJ7YKBBXX      14.57541607   9.4998418  -1.5243075   4.05044314
#> R15696_HJ7YKBBXX     -36.19112106 -11.7878014  -1.9139568   2.43833524
#> R15697_HJ7YKBBXX     -32.97395435 -10.5584959  -5.9951658  -1.64572014
#> R15698_HJ7YKBBXX     -26.63738663 -13.0774971  -3.8686250   3.56467789
#> R15699_HJ7WJBBXX      -6.09006937 -15.8113694  -4.6139460  -2.43074481
#> R15704_HJ7WJBBXX     -11.87414599  -6.3529735   0.1578865  -9.61774420
#> R15705_HJ7WJBBXX       5.04675994  -9.1416051  -6.4863863 -14.38143151
#> R15706_HJ7WJBBXX      -7.57771357 -12.9536659 -24.1913758 -15.44096443
#> R15707_HJ7WJBBXX      22.97912851 -18.6788823 -10.7453562  -7.52874466
#> R15668_HJ7YKBBXX     -22.47363917   9.4818983  -9.0125798  13.48320232
#> R15669_HJ7YKBBXX      -6.30288153  -1.1172437  -9.7615906  18.99595041
#> R15670_HJ7YKBBXX       8.39850049  -4.3365513 -11.9206142  18.83924279
#> R15671_HJ7YKBBXX      18.14801373  -1.6833581 -16.5953092  15.95570042
#> R15672_HJ7YKBBXX     -16.78589401  -3.2907445  -1.4260908  -9.37141174
#> R15673_HJ7YKBBXX      10.80774484  -8.1593904  -7.2104589 -14.33629885
#> R15674_HJ7YKBBXX      32.79676193 -15.5228785  -4.1916429 -14.56973766
#> R15675_HJ7YKBBXX       2.94806312 -13.1576799 -23.8969824 -16.90398708
#> R15676_HJ7YKBBXX      -0.92546253  18.4953698  -1.7215894  -5.59109745
#> R15677_HJ7YKBBXX       6.41688159  23.7614229  -0.8367465  -2.47046326
#> R15678_HJ7YKBBXX      12.33193762  14.1034632  -4.3318075  -0.19513633
#> R15679_HJ7YKBBXX      14.16988631  13.3095111 -10.4439252  -3.92693596
#> R15680_HJ7YKBBXX     -31.38961067  -3.8993395  -6.9767396  -2.64846514
#> R15681_HJ7YKBBXX     -26.01334717  -5.2049767 -11.7218630  -8.34915192
#> R15682_HJ7YKBBXX     -15.86608691  -5.1414662  -7.8854246  -2.89547168
#> R15683_HJ7YKBBXX      -3.66815564  -7.6770113  -9.3409286  -8.82045573
#> R15700_HJ7YKBBXX     -11.02992713  -6.0318263   0.8832234  -4.61794879
#> R15701_HJ7YKBBXX       9.39932980 -10.1057791  -5.7925537  -6.46039718
#> R15702_HJ7YKBBXX       8.14515574 -13.5745553 -14.1833515  -6.95550632
#> R15703_HJ7YKBBXX       8.25892586 -18.7282452 -22.9053357  -7.78622350
#>                               qSV5         qSV6         qSV7         qSV8
#> DLPFC_Br1385_A1_ribo  11.141392889  -3.44302326   2.98386105  -6.50987878
#> DLPFC_Br1385_A2_ribo   4.419143546   0.40922314  -2.37499087  -8.69491776
#> DLPFC_Br1385_A3_ribo   6.704673255   4.04006397  -4.74667441  -3.25590530
#> DLPFC_Br1385_A4_ribo   7.984166581   2.32293295  -2.93055980  -5.97247123
#> DLPFC_Br1729_B1_ribo   8.342837503  -5.17529828   3.77661529  -0.65724024
#> DLPFC_Br1729_B2_ribo   6.394617926  -3.86921687   1.15988987  -0.34774258
#> DLPFC_Br1729_B3_ribo   3.052663627   2.47453394  -1.54778588   4.98135305
#> DLPFC_Br1729_B4_ribo   6.238913839  -3.00444677   3.04937649   2.03544670
#> DLPFC_Br2015_F1_ribo   5.720268347  -5.87400091   1.11533081  -1.49883692
#> DLPFC_Br2015_F2_ribo  12.200529693  -8.64410316   5.37982662   2.93898944
#> DLPFC_Br2015_F3_ribo  -0.118680035   1.22586169  -5.47962809   1.69897483
#> DLPFC_Br2015_F4_ribo   7.157545659  -3.99347871  -0.67439158   3.13492579
#> DLPFC_Br2020_C1_ribo   3.092807104  -5.10967722   5.99859417  -1.17111044
#> DLPFC_Br2020_C2_ribo   3.027293421  -5.67150387   6.91456442   0.73487228
#> DLPFC_Br2020_C3_ribo   1.282431737   4.56743003   0.61168203   7.24053257
#> DLPFC_Br2020_C4_ribo   0.135405566  -1.70550652   4.54534600   4.04130374
#> DLPFC_Br2074_D1_ribo  -2.601640747  -0.95819630   5.31380384  -0.19198154
#> DLPFC_Br2074_D2_ribo   0.830560038  -4.02546858   8.33826001   4.24772457
#> DLPFC_Br2074_D3_ribo   2.203314121   2.19220325   5.03389815  13.04174522
#> DLPFC_Br2074_D4_ribo  -0.399604230  -2.22517322   6.76253987   5.25875651
#> Br2623-H1a            -3.794975553   8.03704931   3.79721321   3.86983130
#> Br2623-H1b            -7.387214059   4.39390100   2.49598267   5.72219214
#> Br2623-H1c            -6.299439151   7.36563253   2.92158117   5.30539973
#> Br2623-H1d            -8.041537398   9.70846385   1.82535797   8.97493612
#> Br5005-H2a            11.394092400  -2.83423640  -7.34104239   1.07249917
#> Br5005-H2b             8.075921628  -9.89586504 -10.05446221   1.65758648
#> Br5005-H2c            10.375216167  -6.68358343  -9.56784974   5.99987166
#> Br5005-H2d            10.354084179  -4.24336203  -7.78563463   4.68935686
#> Br5370-H3a             5.897676394  -2.85907474   5.01496838  -1.03206677
#> Br5370-H3b             2.148627015  -8.19778310   4.71891553  -1.82446547
#> Br5370-H3c             6.496900565  -6.80858400   6.99550322   1.35259399
#> Br5370-H3d             3.328549751  -1.46495173   1.17876950   1.05081585
#> Br5889-H4a             3.755200905  11.35881017   3.25807420  -0.64785274
#> Br5889-H4b             0.425307904   9.95300113   1.46693491  -1.99297533
#> Br5889-H4c            -1.763194083  11.76694528  -0.15059354  -0.72423413
#> Br5889-H4d            -1.449700991  13.77627321  -1.29659712  -0.50075079
#> Br5941-H5a             2.145541388   1.79051244   7.15563594  -6.28408423
#> Br5941-H5b            -6.663763396   1.34640915   5.03945334 -10.76626677
#> Br5941-H5c            -2.745088746   0.36582584   8.11844881  -7.90239796
#> Br5941-H5d            -0.852690352   1.98781700   7.99054504  -6.67818603
#> Br2623-C1a           -15.966030994  -4.86346372   2.40942022   8.34606084
#> Br2623-C1b            -8.733963580  -6.21589342   6.98165799   7.49659828
#> Br2623-C1c           -15.217862232  -5.11223153   2.81765427   9.40794738
#> Br2623-C1d           -15.009328917  -5.48742858   3.74731299   9.99390484
#> Br5005-C2a             1.063087571 -16.34210352 -11.32938966   4.31578682
#> Br5005-C2b             1.165852317 -18.26130406 -13.13063939   4.73882253
#> Br5005-C2c            -3.445838325 -16.22440382 -15.58937939   2.37277447
#> Br5005-C2d            -3.547378714 -17.72957296 -13.85065395   4.06655818
#> Br5370-C3a            -3.475201250 -14.01992954   5.38069218   0.30337030
#> Br5370-C3b            -1.639341975 -17.07260720   7.12023943   1.51086814
#> Br5370-C3c            -5.175811535 -14.90848184   4.86346710   1.90033485
#> Br5370-C3d            -2.911425211 -15.16479198   3.25813197   3.79879662
#> Br5889-C4a            -0.558408582  -2.59430270   7.66608957   1.30641570
#> Br5889-C4b             1.911735066  -3.72307313  10.95024807   1.82394799
#> Br5889-C4c            -2.289105462  -1.46117315   8.22210228   0.99136775
#> Br5889-C4d            -1.357049575  -0.65332825   9.18846698   2.50115503
#> Br5941-C5a           -12.475813004 -12.01879072   7.74177658  -6.65813966
#> Br5941-C5b            -9.533509201 -11.88860163   9.71763317  -7.96242015
#> Br5941-C5c           -14.789651619 -12.15652895   6.20071765  -7.77665665
#> Br5941-C5d           -14.474827356 -11.29716930   7.17018247  -6.52898166
#> Br2623-M1a           -18.918623386   6.98545395  -5.20108357   1.70770075
#> Br2623-M1b           -20.144838377   5.83854041  -3.63854179   2.05479598
#> Br2623-M1c           -17.981448623   6.72496704  -5.32142608   3.04692424
#> Br2623-M1d           -23.354617478   7.69983471  -7.09776960   1.28516940
#> Br5005-M2a            -2.619658780  -7.73375480 -17.96959466   0.17784506
#> Br5005-M2b            -5.944389682  -4.91064619 -20.88200708  -2.00836299
#> Br5005-M2d            -7.346491902  -4.88900523 -22.15989815   0.62976796
#> Br5370-M3a            -6.039365088  -6.26773643  -1.77528955  -5.44936248
#> Br5370-M3b            -2.905847052  -7.97260575   0.60954048  -2.80628686
#> Br5370-M3c            -6.529070064  -7.28635083  -2.57988503  -3.44843820
#> Br5370-M3d            -3.907209741  -7.60241778  -1.76660053  -1.49105002
#> Br5889-M4a            -7.851630779   8.41755451  -3.92232569  -6.81172885
#> Br5889-M4b            -7.591072204   8.60103040  -2.87816126  -4.56042939
#> Br5889-M4c            -7.411379581   9.31442697  -3.78378265  -3.91202702
#> Br5889-M4d           -11.605949970  13.56073723  -6.53557059  -4.31280838
#> Br5941-M5a           -11.673420058  -1.97662485   1.20375703 -12.34746178
#> Br5941-M5b           -13.755985790  -0.99021802   0.24693987 -12.82217285
#> Br5941-M5c           -12.337757558  -0.69665827   1.50009942 -13.08605454
#> Br5941-M5d           -17.756452999   1.89054923  -1.38550771 -13.54121744
#> R15684_HJ7YKBBXX       6.084665033   2.78989014   2.53135676  -1.29806887
#> R15685_HJ7YKBBXX       2.740734633   1.55658299   0.68400274  -1.79732788
#> R15686_HJ7YKBBXX       0.006488189   1.62544557  -0.70675865  -2.39164867
#> R15687_HJ7YKBBXX      -0.266254929   2.63231332  -2.59068382   1.00984340
#> R15688_HJ7YKBBXX       8.481483830  -1.87443254  -9.02965679   0.56910197
#> R15689_HJ7YKBBXX      11.316004479  -2.05550078  -7.65282625   1.79285140
#> R15690_HJ7YKBBXX      11.460952608  -3.33404792  -9.02539184   4.20537993
#> R15691_HJ7YKBBXX       5.099674427   0.55142807 -14.89991456   2.75480218
#> R15692_HJ7YKBBXX       4.828856389  11.55587417   2.42655628  -3.09774967
#> R15693_HJ7YKBBXX       2.099018973  12.91681497  -0.10536350  -1.54153919
#> R15694_HJ7YKBBXX       1.520107057  12.46846146   0.17652280  -1.17014036
#> R15695_HJ7YKBBXX       2.285642824  18.29016947  -3.86521473   2.04390350
#> R15696_HJ7YKBBXX      -8.257223260  10.63243505  -1.96093141   4.85725557
#> R15697_HJ7YKBBXX      -4.138920439   8.55965994   0.05110657   6.64324024
#> R15698_HJ7YKBBXX     -10.035438124  12.49590275  -1.96251430   3.83854859
#> R15699_HJ7WJBBXX      -5.680468669   9.69744036  -0.45415007   9.77207727
#> R15704_HJ7WJBBXX      20.257759783  -0.67165572   6.09161879  -3.83308694
#> R15705_HJ7WJBBXX      18.532581969  -2.60315194   3.53166799  -0.06523608
#> R15706_HJ7WJBBXX      17.374291296   2.72775307  -0.07534268  -0.82169162
#> R15707_HJ7WJBBXX       8.577156321   9.05140778  -7.80610462  -5.82709900
#> R15668_HJ7YKBBXX      10.326604526  -5.20079227   9.45027067  -0.83874495
#> R15669_HJ7YKBBXX       4.711290154  -0.06242273   3.33287437  -1.37173645
#> R15670_HJ7YKBBXX       2.677915995   0.49039069   1.59458437  -0.75249778
#> R15671_HJ7YKBBXX       6.937024214  -0.81759605   3.85354327   1.42580135
#> R15672_HJ7YKBBXX      12.860401113  -1.06235299   0.90849573  -4.56383788
#> R15673_HJ7YKBBXX      13.246307541  -2.66733220  -0.57771715  -3.67794141
#> R15674_HJ7YKBBXX      16.984243746  -0.17828049  -0.46412195  -2.14055000
#> R15675_HJ7YKBBXX      11.640498562   2.34801780  -4.16639658  -2.48223104
#> R15676_HJ7YKBBXX      11.706164742   7.66288424   5.76152017   4.86993702
#> R15677_HJ7YKBBXX       8.976959273  10.46867619   5.86056405   2.67413029
#> R15678_HJ7YKBBXX       6.062919153  14.06202060   0.91534140   3.44430142
#> R15679_HJ7YKBBXX       5.956131299  13.28116841  -0.80546616   4.47390183
#> R15680_HJ7YKBBXX      -1.266492616   6.97799919   5.28616414   5.17173861
#> R15681_HJ7YKBBXX       1.163889697   1.30866312   7.10302272   9.25070927
#> R15682_HJ7YKBBXX      -6.139051049   5.40834972   4.06115240   7.08415166
#> R15683_HJ7YKBBXX      -7.881156390   0.85471730   4.79497212  10.22645114
#> R15700_HJ7YKBBXX       8.581697390   3.46635049  -1.08670439  -7.66039383
#> R15701_HJ7YKBBXX       6.988874944   5.44369102  -2.65567450  -7.37479870
#> R15702_HJ7YKBBXX       7.541292395   7.96755281  -5.46250425  -6.54096966
#> R15703_HJ7YKBBXX       6.564300203   9.32725288  -8.27128477  -7.51049382
#>                             qSV9        qSV10
#> DLPFC_Br1385_A1_ribo   9.8243688   2.87348736
#> DLPFC_Br1385_A2_ribo  11.0340074  -4.51978473
#> DLPFC_Br1385_A3_ribo   4.6709164  -0.17386295
#> DLPFC_Br1385_A4_ribo   6.4517439   2.86944690
#> DLPFC_Br1729_B1_ribo   3.0581727  -1.79807933
#> DLPFC_Br1729_B2_ribo   3.2788271  -1.34563842
#> DLPFC_Br1729_B3_ribo  -3.1190965  -9.11641464
#> DLPFC_Br1729_B4_ribo  -1.7244255   3.24386626
#> DLPFC_Br2015_F1_ribo   4.8192959  -9.44602281
#> DLPFC_Br2015_F2_ribo  -1.5428396   1.52952354
#> DLPFC_Br2015_F3_ribo  -3.3592985  -2.12430075
#> DLPFC_Br2015_F4_ribo   0.2072023  -6.43366614
#> DLPFC_Br2020_C1_ribo  -0.8082786   0.45578732
#> DLPFC_Br2020_C2_ribo  -4.5753279   5.75021053
#> DLPFC_Br2020_C3_ribo  -9.7326814  -8.82489308
#> DLPFC_Br2020_C4_ribo  -6.3680560   6.28248400
#> DLPFC_Br2074_D1_ribo   6.3136482 -13.26474597
#> DLPFC_Br2074_D2_ribo   1.0663707  -1.00189420
#> DLPFC_Br2074_D3_ribo  -4.8492546 -10.74889404
#> DLPFC_Br2074_D4_ribo  -0.9058119   0.30861052
#> Br2623-H1a            -1.5391316   3.02821912
#> Br2623-H1b            -1.4885192   7.42740947
#> Br2623-H1c            -4.1156984   7.03574965
#> Br2623-H1d            -5.7731920   4.62329366
#> Br5005-H2a            -7.9693215  -0.94895445
#> Br5005-H2b            -6.7044238   6.27789702
#> Br5005-H2c            -6.5747764   4.76108149
#> Br5005-H2d           -11.1266072   2.66555891
#> Br5370-H3a            -0.8137142   2.95764800
#> Br5370-H3b            -0.2375875   8.57224288
#> Br5370-H3c            -3.0934358   7.57770186
#> Br5370-H3d            -2.3743215   1.71642126
#> Br5889-H4a            -3.8565310   1.81772428
#> Br5889-H4b            -1.9687045   3.14463974
#> Br5889-H4c            -2.8529300   2.54074891
#> Br5889-H4d            -3.5072908  -2.01117823
#> Br5941-H5a           -10.1536555  -3.45237434
#> Br5941-H5b            -8.9025022  -2.43049216
#> Br5941-H5c           -13.1693596   4.34166479
#> Br5941-H5d           -13.3171369  -0.49696519
#> Br2623-C1a             4.8240943   0.35799084
#> Br2623-C1b             4.2897530   0.42680097
#> Br2623-C1c             5.0882370  -0.78650185
#> Br2623-C1d             4.1155526   0.84889863
#> Br5005-C2a            -0.8783784  -2.29575028
#> Br5005-C2b            -1.2532074  -1.71436410
#> Br5005-C2c             0.4331141  -3.84775311
#> Br5005-C2d            -1.1725310  -1.67813891
#> Br5370-C3a             5.9412536  -5.70238747
#> Br5370-C3b             5.8960383  -2.91111410
#> Br5370-C3c             5.4279049  -5.00136503
#> Br5370-C3d             5.7505195  -5.17658703
#> Br5889-C4a             4.8495742  -1.44077548
#> Br5889-C4b             5.5977048   1.40110784
#> Br5889-C4c             5.7187497  -1.05124882
#> Br5889-C4d             3.9557010  -0.26851649
#> Br5941-C5a            -6.8784316  -3.30435915
#> Br5941-C5b            -5.5894491  -2.77675849
#> Br5941-C5c            -5.5016341  -4.41685552
#> Br5941-C5d            -7.0063113  -3.52977893
#> Br2623-M1a             5.9953528   0.84967051
#> Br2623-M1b             5.8751831   3.85446465
#> Br2623-M1c             5.7468301   4.23770098
#> Br2623-M1d             5.5461401   5.00287046
#> Br5005-M2a            -0.6583740   0.70741235
#> Br5005-M2b             1.5670544  -1.35575965
#> Br5005-M2d            -2.1174174   4.89743192
#> Br5370-M3a             6.3596238   1.98724090
#> Br5370-M3b             3.4194122   5.80627791
#> Br5370-M3c             4.8264228   4.40407027
#> Br5370-M3d             2.5119176   4.75145608
#> Br5889-M4a             5.8746743  -2.00244320
#> Br5889-M4b             4.6333769   1.86397382
#> Br5889-M4c             2.5104770   3.47160914
#> Br5889-M4d             2.6778377  -1.13556145
#> Br5941-M5a            -5.4394327  -0.05999455
#> Br5941-M5b            -4.5573888  -0.80582039
#> Br5941-M5c            -6.6706313   1.08324916
#> Br5941-M5d            -5.7696835  -0.09889846
#> R15684_HJ7YKBBXX       1.3908037  -4.38534935
#> R15685_HJ7YKBBXX       1.8364611  -1.08731529
#> R15686_HJ7YKBBXX       0.7038484   0.97279040
#> R15687_HJ7YKBBXX      -1.9233792  -1.34222543
#> R15688_HJ7YKBBXX      -3.3162482  -2.90187327
#> R15689_HJ7YKBBXX      -5.1534583  -1.89638673
#> R15690_HJ7YKBBXX      -6.1987264  -1.91172625
#> R15691_HJ7YKBBXX      -8.5808740   1.13057321
#> R15692_HJ7YKBBXX       2.8309043  -7.12237204
#> R15693_HJ7YKBBXX       1.0072254  -0.67334064
#> R15694_HJ7YKBBXX       1.1105032  -1.34906604
#> R15695_HJ7YKBBXX      -2.8325732  -6.71495171
#> R15696_HJ7YKBBXX       3.4774391  -1.16915257
#> R15697_HJ7YKBBXX       2.3618532   0.20341767
#> R15698_HJ7YKBBXX       1.9872265  -0.69745254
#> R15699_HJ7WJBBXX       0.2826486   3.46537491
#> R15704_HJ7WJBBXX       5.1617071   3.42476737
#> R15705_HJ7WJBBXX       3.3753727   4.83888592
#> R15706_HJ7WJBBXX       5.3993329  -5.84144058
#> R15707_HJ7WJBBXX       3.0761587  -2.75477854
#> R15668_HJ7YKBBXX       1.8115732   2.52221938
#> R15669_HJ7YKBBXX      -0.2985483  -0.29609014
#> R15670_HJ7YKBBXX      -1.4168288  -0.54443888
#> R15671_HJ7YKBBXX      -2.4570293  -1.08776646
#> R15672_HJ7YKBBXX       6.1304823   5.26964231
#> R15673_HJ7YKBBXX       4.0995298   7.53483906
#> R15674_HJ7YKBBXX       4.1269084  11.87348169
#> R15675_HJ7YKBBXX       5.4695546  -3.26671036
#> R15676_HJ7YKBBXX      -2.9531894  -0.09468893
#> R15677_HJ7YKBBXX      -1.8816061   0.62181301
#> R15678_HJ7YKBBXX      -4.1129555  -4.43275800
#> R15679_HJ7YKBBXX      -2.9938276  -5.11602098
#> R15680_HJ7YKBBXX      -0.1813847   0.66656447
#> R15681_HJ7YKBBXX       0.6091194   2.17760532
#> R15682_HJ7YKBBXX       0.7228375   2.00366849
#> R15683_HJ7YKBBXX      -0.6550396   2.36003039
#> R15700_HJ7YKBBXX       4.5795722   3.99600859
#> R15701_HJ7YKBBXX       6.2333895   2.65152820
#> R15702_HJ7YKBBXX       3.0890056  -2.37396701
#> R15703_HJ7YKBBXX       3.9419382  -6.90808869